Data.Bits.Coded

Description

Documentation

class Coded c where

Unaligned codes

Methods

encode :: MonadPut m => c -> Coding m ()

encodeMany :: (MonadPut m, Foldable t) => t c -> Coding m ()

decode :: MonadGet m => Coding m c

newtype Unary n

Unary-coded integers

>>> runPutL . runEncode $ encode (Unary 1) >> flush
"\128"
>>> runPutL . runEncode $ encode (Unary 7) >> flush
"\254"

Constructors

Unary

Fields

unUnary :: n

newtype Elias c n

Representation for Elias Gamma and Delta codes. A positive integer n is encoded by encoding the position of its most significant bit, and then the binary representation of the rest of the number.

Constructors

Elias

Fields

unElias :: n

type Gamma c n = Elias (Unary c) n

Elias Gamma codes the position of the most significant in Unary.

type Delta c n = Elias (Gamma c n) n

Elias Delta codes the position of the most significant bit in Elias Gamma.

runEncode :: MonadPut m => Coding m () -> m ()

runDecode :: MonadGet m => Coding m a -> m a