Data.Bits.Extras

Description

Calculate a number of fiddly bit operations using fast de Bruijn multiplication tables.

Documentation

class (Num t, Bits t) => Ranked t where

Methods

lsb :: t -> Int

Calculate the least significant set bit using a debruijn multiplication table. NB: The result of this function is undefined when given 0.

rank :: t -> Int

Calculate the number of trailing 0 bits.

nlz :: t -> Int

Calculate the number of leading zeros.

log2 :: Word32 -> Int

msb :: Ranked t => t -> Int

Calculate the most significant set bit.

w8 :: Integral a => a -> Word8

w16 :: Integral a => a -> Word16

w32 :: Integral a => a -> Word32

w64 :: Integral a => a -> Word64

assignBit :: Bits b => b -> Int -> Bool -> b

zeroBits :: Bits b => b

oneBits :: Bits b => b

srl :: Bits b => b -> Int -> b

Shift Right Logical (i.e., without sign extension)

NB: When used on negative Integers, hilarity may ensue.