Data.Bits.Coding

Description

Documentation

newtype Coding m a

Constructors

Coding

Fields

runCoding :: forall r. (a -> Int -> Word8 -> m r) -> Int -> Word8 -> m r

Get

getAligned :: MonadGet m => m a -> Coding m a

Get something from byte-aligned storage, starting on the next byte and flushing any left over bits.

getBit :: MonadGet m => Coding m Bool

Get a single bit, consuming an entire byte if the bit buffer is empty

Put

putAligned :: MonadPut m => m a -> Coding m a

Emit any contents from the bit buffer.

putBit :: MonadPut m => Bool -> Coding m ()

Put a single bit, emitting an entire byte if the bit buffer is full